chore: rename the directory
(cherry picked from commit cfa4ebc4acd93d9dee4512f3e80763521dfccaac)
