+++ Source/wunthshin/Subsystem/CharacterSubsystem/CharacterSubsystem.cpp
@@ -4,7 +4,7 @@
 #include "CharacterSubsystem.h"
 
 #include "wunthshin/Components/CharacterStats/CharacterStatsComponent.h"
-#include "wunthshin/Data/CharacterTableRow/CharacterTableRow.h"
+#include "wunthshin/Data/Characters/CharacterTableRow/CharacterTableRow.h"
 
 UCharacterSubsystem::UCharacterSubsystem()
 {
